refactor(pipnn): move distance blocks to topk
