+++ diskann/src/graph/pipnn/simd.rs
@@ -43,48 +43,3 @@ where
 {
     type Vector = DefaultVector<A>;
 }
-
-/// One group of distances and their column indexes in the supplied slice.
-#[derive(Clone, Copy)]
-pub(super) enum DistanceBlock<'a, F: PiPNNSIMDVector> {
-    Simd {
-        first_idx: usize,
-        values: F,
-        lanes: &'a [f32],
-    },
-    Scalar {
-        idx: usize,
-        distance: f32,
-    },
-}
-
-/// Iterate over distance groups without a callback on the ranking hot path.
-///
-/// Scalar lanes borrow the input slice, so sharing a block across updates needs
-/// no temporary array. Tail entries follow in column order. The caller keeps the
-/// selected architecture's execution scope active.
-#[inline(always)]
-pub(super) fn distance_blocks<A: PiPNNSIMDSchema>(
-    arch: A,
-    distances: &[f32],
-) -> impl Iterator<Item = DistanceBlock<'_, A::Vector>> {
-    let simd_end = distances.len() - distances.len() % A::Vector::LANES;
-    distances[..simd_end]
-        .chunks_exact(A::Vector::LANES)
-        .enumerate()
-        .map(move |(group, lanes)| DistanceBlock::Simd {
-            first_idx: group * A::Vector::LANES,
-            // SAFETY: chunks_exact yields one complete SIMD group.
-            values: unsafe { A::Vector::load_simd(arch, lanes.as_ptr()) },
-            lanes,
-        })
-        .chain(
-            distances[simd_end..]
-                .iter()
-                .enumerate()
-                .map(move |(tail, &distance)| DistanceBlock::Scalar {
-                    idx: simd_end + tail,
-                    distance,
-                }),
-        )
-}
